% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCusum.R
\name{cusumData}
\alias{cusumData}
\title{The dataset 'cusumData' in cusumActiveManager package}
\description{
Data in the example is an xts object containing two monthly returns in each column.
The first column is the fund returns and the second column is the
benchmark returns. The data is from Jan 2005 to Dec 2003.
}
