% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robRiskBudget.R
\name{robRiskBudget}
\alias{robRiskBudget}
\title{Simple and Robust Risk Budgeting with Expected Shortfall}
\usage{
robRiskBudget(returns, rf = 0, ER = NULL, IR = NULL, TE = NULL,
  corMat = NULL, ES = NULL, ESMethod = c("modified", "gaussian",
  "historical"), corMatMethod = c("auto", "mcd", "pairwiseQC", "pairwiseGK"),
  targetVol = 0.02, shrink = FALSE, avgCor = FALSE, p = 0.95)
}
\arguments{
\item{returns}{matrix with a time series of returns for each strategy}

\item{rf}{a vector of the risk free rates in each time period, but could also be the returns of a common benchmark}

\item{ER}{a vector of exogeneously derived prospective expeted returns. Either IR or ER must be specified.}

\item{IR}{a vector of exogeneously derived prospective information ratios. Either IR or ER must be specified.}

\item{TE}{a vector of exogneously derived prospective volatilites (tracking errors). If it is not specified, it will be estimated from the time series of asset / strategy returns.}

\item{corMat}{Exogneously derived correlation matrix. If it is not specified, it will be estimated from the time series of asset / strategy returns.}

\item{ES}{the user's speculated expected shortfall. If it is not specified, it will be estimated from the time series of asset / strategy returns.}

\item{ESMethod}{the method used to compute ES from PerformanceAnalytics}

\item{corMatMethod}{the method used to estimate the robust correlation matrix, default = "auto"}

\item{targetVol}{the target volatility}

\item{shrink}{logical value that determines whether or not we want to shrink the expected returns to their grand mean}

\item{avgCor}{logical value that determines if we want to set each off-diagonal element to the average of all the off-diagonal elements}

\item{p}{specified probability used to compute ES}
}
\value{
\code{robRiskBudget} returns a \code{list} containing the following objects:
\item{riskBudget}{The vector of final risk budgets in each asset after the optimization}
\item{excessReturns}{The achieved volatility from the risk budgets}
\item{avgOffDiag}{The difference between the achieved volatility and the targeted volatility}
\item{targetVol}{The matrix of risk budgets in each iteration columnwise}
\item{avgCor}{The matrix whose columns represent the vector of normalized distances in each iteration}
\item{TE}{The vector of tracking errors}
\item{ES}{The vector of estimated ES}
\item{modIR}{The vector od the modified IR by the fraction TEES}
}
\description{
We implement the tracking error allocation in both correlated and noncorrelated
alpha sources
}
\details{
(TBA)
}
\examples{
data("RussellData")
RussellData = data
rf = RussellData[, 16]
robRiskData = RussellData[, 1:15]

riskBudget = robRiskBudget(robRiskData, rf = rf, shrink = TRUE, avgCor = TRUE,
ESMethod = "historical", corMatMethod = "mcd")
robRiskBudget(robRiskData, shrink = TRUE, corMatMethod = "pairwiseQC", avgCor = FALSE)
robRiskBudget(robRiskData, shrink = TRUE, corMatMethod = "mcd", avgCor = TRUE)

}
\references{
Thomas Philips and Michael Liu. "Simple and Robust Risk Budgeting with Expected Shortfall",
The Journal of Portfolio Management, Fall 2011, pp. 1-13.
}
\author{
Thomas Philips, Chindhanai Uthaisaad
}
