% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repCusum.R
\name{summary.cusumActMgr}
\alias{summary.cusumActMgr}
\title{Summarizing a cusumActMgr object}
\usage{
\method{summary}{cusumActMgr}(object, digits = 3, ...)
}
\arguments{
\item{object}{an object of class \code{cusumActMgr} returned by \code{cusumActMgr}.}

\item{digits}{number of significants digits to use when printing.
Default is 3.}

\item{...}{futher arguments passed to or from other methods.}
}
\value{
Returns an object of class \code{summary.ffm}.

Object of class \code{summary.cusumActMgr} is a list of length 10 containing:
\item{Logarithmic Excess Returns}{}
\item{Annualized Moving Average}{The annualized moving average of the 
logarithmic excess returns}
\item{Tracking Error}{}
\item{Information Ratios}{}
\item{Lindley's Recursion}{}
\item{Annualized Cusum IR}{}
\item{Annualized Cusum Excess Return}{}
\item{Excess Volatility}{Excess volatility of the fund, the benchmark and the excess return}
\item{Summary Annualized cusumIR}{}
\item{Summary Annualized Cusum Excess Returns}{}
}
\description{
\code{summary} method for object of class \code{cusumActMgr}.
Returned object is of class {summary.cusumActMgr}.
}
\examples{
data(cusumData)
results = cusumActMgr(portfolioName = "Parvest", benchmarkName = "RUS2500",
data = cusumData)
summary.cusumActMgr(results)

}
\author{
Chindhanai Uthaisaad.
}
