% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartCusum.R
\name{chartCusum}
\alias{chartCusum}
\title{cusumActMgr Plots}
\usage{
chartCusum(object, digits = 3, which = NULL, ...)
}
\arguments{
\item{object}{An object of class cusumActMgr returned by cusumActMgr.}

\item{digits}{The number of digits of numerical values in graphs}

\item{which}{a number or a vector of numbers to indicate the type of plots.
If a subset of the plots is required, specify a subset of the numbers 1:8
for plots. The numbers 1 through 8 represent: \cr \cr
1 = Barplot of log-excess returns with annually moving average returns, \cr
2 = Plot of tracking error, \cr
3 = Barplot of the information ratio, \cr
4 = Plot of cusumIR with protractors of slopes representing IRs, \cr
5 = Plot of Lindley's Recursion with the thresholds, \cr
6 = Plot of excess volatility, \cr
7 = Scatter plot between the fund and benchmark returns with robust regression \cr
8 = Plot of cusum for returns with protractors of slopes representing the annualized returns. \cr}

\item{...}{other graphics parameters in plot}
}
\value{
Graph(s) as specified by the user
}
\description{
Plot the elements of an \code{cusumActMgr} object.
}
\examples{
data(cusumData)
results = cusumActMgr(portfolioName = "Parvest", benchmarkName = "RUS2500",
data = cusumData)
chartCusum(results, which = 8)
chartCusum(results, which = c(1,3,4,7))
}
\author{
Chindhanai Uthaisaad
}
