% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robRiskBudget.R
\name{robRiskBudget}
\alias{robRiskBudget}
\title{Simple and Robust Risk Budgeting with Expected Shortfall}
\usage{
robRiskBudget(returns = NULL, rf = 0, ER = NULL, IR = NULL, TE = NULL,
  corMat = NULL, ES = NULL, ESMethod = c("modified", "gaussian",
  "historical"), corMatMethod = c("auto", "mcd", "pairwiseQC", "pairwiseGK"),
  targetVol = 0.02, shrink = FALSE, avgCor = FALSE, p = 0.95,
  lower = 0, upper = 1, K = 100, maxit = 50, tol = 1e-05)
}
\arguments{
\item{returns}{A matrix with a time series of returns for each asset / strategy}

\item{rf}{A vector with the risk free rate in each time period; could also be the returns of a common benchmark}

\item{ER}{A vector of exogeneously derived prospective expected returns. Either IR or ER must be specified.}

\item{IR}{A vector of exogeneously derived prospective information ratios. Either IR or ER must be specified.}

\item{TE}{A vector of exogneously derived prospective volatilites (tracking errors). If it is not specified,
it will be estimated from the time series of asset / strategy returns.}

\item{corMat}{Exogneously derived correlation matrix. If it is not specified, it will be estimated
from the time series of asset / strategy returns.}

\item{ES}{The user's speculated expected shortfall. If it is not specified, it will be estimated
from the time series of asset / strategy returns using one of 3 user-selectable methods.}

\item{ESMethod}{The method used to compute ES (the methods are available in PerformanceAnalytics)}

\item{corMatMethod}{the method used to estimate the robust correlation matrix,
(the methods are available in library robust, and the default is "auto")}

\item{targetVol}{The target volatility}

\item{shrink}{Logical value that determines whether or not we want to shrink the expected returns toward their grand mean
Uses James-Stein shrinkage, but is applicable only when expected returns are estimated by averaging historical returns,}

\item{avgCor}{Logical value that determines if we want to set each off-diagonal element in the correlation matrix
to the average of all its off-diagonal elements}

\item{p}{User-specified confidence level for computing ES}

\item{lower}{A numeric or vector of lower bounds \deqn{\vec{\sigma}^L} of risk budgets. Default is 0.}

\item{upper}{A numeric or vector of upper bounds \deqn{\vec{\sigma}^U} of risk budgets. Default is 1.}

\item{K}{A tuning factor for the iterative algorithm. Default is 100}

\item{maxit}{A number of maximum iterations. Default is 50.}

\item{tol}{An accuracy of the estimation with respect to the targeted volatility. Default is {10^{-5}}}
}
\value{
\code{robRiskBudget} returns a \code{list} containing the following objects:
\item{initialRiskBudget}{The vector of unconstrained risk budgets}
\item{finalRiskBudget}{The vector of final risk budgets for each asset}
\item{iterativeRiskBudget}{The matrix of risk budgets in each iteration columnwise}
\item{targetVol}{The target volatility for the portfolio}
\item{avgCor}{The average of all off-diagonal correlations}
\item{ER}{The average returns (shrunk if called for) or the user supplied expected returns}
\item{TE}{The estimated tracking errors or the user supplied tracking errors}
\item{ES}{The vector of estimated ESs or the user supplied ESs}
\item{IR}{The shrunk average IRs or the user supplied IRs}
\item{modIR}{The vector of modified IRs}
}
\details{
This function starts by computing an unconstrained meean variance optimmal portfolio in
which the risk budgets are expressed as a function of the various information ratios
(Information Ratio = Excess Return / Volatility (Excess Return) ). This can be solved in closed form.
We then replace each Information Ratio with its corresponding Modified Information Ratio
(Modified Information Ratio = Excess Return / Expected Shortfall (Excess Return) ).
In the special case when all the returns are drawn from the same scale family (e.g. Gaussian),
ES is just a constant multiple of \deqn{\sigma}, and the solution is unchanged. But if some strategies have
a higher level of tail risk relative to others, their risk budget will decline. In effect, risk
is allocated away from strategies with high tail risk and to strategies with low tail risk.
}
\examples{
data("RussellData")
RussellData = data
rf = RussellData[, 16]
robRiskData = RussellData[, 1:15]

riskBudget = robRiskBudget(robRiskData, rf = rf, shrink = TRUE, avgCor = TRUE,
ESMethod = "historical", corMatMethod = "mcd")
robRiskBudget(robRiskData, shrink = TRUE, corMatMethod = "pairwiseQC", avgCor = TRUE)
robRiskBudget(robRiskData, shrink = TRUE, corMatMethod = "mcd", avgCor = TRUE, upper = 0.0123)

}
\references{
Thomas Philips and Michael Liu. "Simple and Robust Risk Budgeting with Expected Shortfall",
The Journal of Portfolio Management, Fall 2011, pp. 1-13.
}
\author{
Thomas Philips, Chindhanai Uthaisaad
}
