% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateARL.R
\name{simulateARL}
\alias{simulateARL}
\title{Simulation for thresholds of the Lindley's recursion}
\usage{
simulateARL(mu, Threshold, delta, EW_constant = 0.9, Fixed_Sigma = 1)
}
\arguments{
\item{mu}{A numeric value representing the information ratio we want to
simulate thresholds for. No default value is set.}

\item{Threshold}{A numeric value representing the threshold for the Lindley's
recursion to be updated in the recursion. No default value is set.}

\item{delta}{A numeric value representing the simulation accuracy.
In other words, we will simulate until 3 * sigma / mu < precision.}

\item{EW_constant}{A numeric value representing the ratio between the former sigma
and the new sigma. The default is set to 0.9}

\item{Fixed_Sigma}{The logical value representing if the sigma should be constant of not.
0 represents fixed sigma, 1 represents weighted sigma. The default is set to 1.}
}
\value{
\code{simulateARL} returns a \code{vector} of the following:
\item{ARL}{The average return length}
\item{s}{The standard deviation of the ARLs}
}
\description{
This function simulates thresholds of Lindley's recursion
used in the function \code{cusumActMgr}.
}
\details{

}
\references{
Philips, T. K., Yashchin, E. and Stein, D. M. (2003). “Using Statistical
Process Control to Monitor Active Managers”, Journal of Portfolio Management,
Fall 2003, pp. 86-94.
}
\author{
Thomas Philips, Chindhanai Uthaisaad.
}
